[Refactor] recent, interest view
관심 매장, 최근 본 매장 약 수정
