+++ app/src/main/java/org/cazait/cazait_android/ui/view/viewmore/ViewMoreFragment.kt
@@ -1,11 +1,13 @@
 package org.cazait.cazait_android.ui.view.viewmore
 
 import androidx.fragment.app.viewModels
+import dagger.hilt.android.AndroidEntryPoint
 import org.cazait.cazait_android.R
 import org.cazait.cazait_android.databinding.FragmentViewMoreBinding
 import org.cazait.cazait_android.ui.base.BaseFragment
 import org.cazait.cazait_android.ui.base.BaseViewModel
 
+@AndroidEntryPoint
 class ViewMoreFragment : BaseFragment<FragmentViewMoreBinding, BaseViewModel>() {
     override val layoutResourceId: Int
         get() = R.layout.fragment_view_more
