+++ app/src/main/res/layout/activity_cafe_rating_review_edit.xml
@@ -104,7 +104,5 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"/>
 
-
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
