+++ app/src/main/res/layout/activity_cafe_recent.xml
@@ -7,7 +7,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/cazait_foam_white"
-        tools:context=".ui.view.recently.CafeRecentlyActivity">
+        tools:context=".ui.view.mypage.recent.CafeRecentActivity">
 
         <View
             android:id="@+id/view_cafe_recently_divide"
@@ -58,7 +58,5 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/view_cafe_recently_divide" />
-
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
