+++ app/src/main/res/layout/activity_login.xml
@@ -88,7 +88,6 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/et_login_user_name" />
 
-
         <TextView
             android:id="@+id/tv_login_sign_up"
             android:layout_width="wrap_content"
