+++ app/src/main/res/layout/fragment_cafe_interest.xml
@@ -20,7 +20,7 @@
             app:layout_constraintTop_toTopOf="parent" />
 
         <TextView
-            android:id="@+id/tv_cafe_like_fragment_shop"
+            android:id="@+id/tv_cafe_interest_shop"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="18dp"
