+++ app/src/main/res/layout/item_cafe_img.xml
@@ -8,6 +8,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:scaleType="fitXY"
-        android:src="@drawable/image_cafe_ex1"/>
+        android:src="@drawable/image_cafe_ex1"
+        android:contentDescription="@null" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
