+++ app/src/main/res/layout/item_cafe_interest.xml
@@ -5,9 +5,8 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-
         <variable
-            name="cafeState"
+            name="cafe"
             type="org.cazait.cazait_android.data.model.Cafe" />
     </data>
 
@@ -52,7 +51,7 @@
                     android:layout_marginBottom="10dp"
                     android:fontFamily="@font/roboto_bold"
                     android:maxLines="2"
-                    android:text="@{cafeState.name, default=카페}"
+                    android:text="@{cafe.name, default=카페}"
                     android:textColor="@color/cazait_black"
                     android:textSize="14sp"
                     app:layout_constraintBottom_toTopOf="@+id/tv_cafe_interest_address"
@@ -68,7 +67,7 @@
                     android:ellipsize="end"
                     android:fontFamily="@font/roboto_regular"
                     android:maxLines="2"
-                    android:text="@{cafeState.address, default=addressaddressaddress}"
+                    android:text="@{cafe.address, default=addressaddressaddress}"
                     android:textColor="@color/cazait_black"
                     android:textSize="12sp"
                     app:layout_constraintBottom_toTopOf="@+id/tv_interest_state"
@@ -86,10 +85,11 @@
                     android:fontFamily="@font/roboto_bold"
                     android:gravity="center"
                     android:minHeight="38dp"
-                    android:text="@{cafeState.state, default=보통}"
+                    android:text="@{cafe.state, default=보통}"
                     android:textAlignment="center"
                     android:textColor="@color/cafe_list_item_normal"
                     android:textSize="20sp"
+                    app:stateColor="@{cafe.state}"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent" />
